Propagate linker flags for kqueue and pthread_workqueue
If libdispatch is using functions found in libkqueue or
libpthread_workqueue, then we want libdispatch.so to
depend on those libraries so we don't have to specify
them explictly in the application linking command.
This change does that by adding KQUEUE_LIBS and
PTHREAD_WORKQUEUE_LIBS to the linking command for libdispatch.la
and exposing HAVE_PTHREAD_WORKQUEUE as a variable to automake.

Signed-off-by: Daniel A. Steffen <dsteffen@apple.com>