sem_timedwait on Linux needs epoch-based timeout
sem_timedwait expects a timeout argument based on epoch time;
the code was converting from dispatch_time to a relative timeout,
but failing to further convert that relative time to an
epoch-based absolute time before calling sem_timedwait.

This fix localizes the change, but is not very efficient.
Arguably it would be better to compute a conversion from
dispatch_time to epoch time during initialization and simply
apply that instead of recomputing it every time we need it.