libpwq as git submodule
On Linux we need recent updates to libpthread_workqueue
to support 6 priority levels instead of 4.  To ensure
we get a compatible libpwq version, use git submodules
to get the right source version, change the autotools
to build it as a substep of building libdispatch, and
statically link the libpwq object files into libdispatch.so
to avoid picking up a system libpthread_workqueue.so via
dynamic linking.

The pieces of this commit are:
 (1) update INSTALL instrictions.
 (2) update configure.ac, Makefile.am to support build
 (3) bind submodule to dgrove-oss/libpwq until pending
     pull requests are merged upsteam into mheily/libpwq
     and we are satisfied that we will not need to make
     further adjustments to libpwq.

Signed-off-by: Daniel A. Steffen <dsteffen@apple.com>