Initial integration of Swift overlay into libdispatch build
This commit adds a src/swift directory that contains
Dispatch.swift and Dispatch.c (based on existing files in
swift/stdlib/public/SDK/Dispatch) and a module.map.in template.
We made some initial changes to Dispatch.swift and Dispatch.c to
adapt them to Linux by dropping MACH API functions/adapters.

The libdispatch build process is extended to include Dispatch.o
in libdispatch.so and to generate Dispatch.swiftmodule,
Dispatch.swiftdoc and a module.map file when configure is
given a --with-swift-toolchain argument.

Swift programs that import Dispatch can be compiled, linked,
and run against the generated files. However, because of
the use of @convention(block) to bridge from Swift blocks to
the -fblocks expected by dispatch, swiftc must be given the
extra command line arguments  -Xcc -fblocks