if building Swift overlay, install into Swift toolchain dir structure
Simplify the build/install logic for the Swift overlay portion of
libdispatch to assume that if the Swift overlay is being built
then the intended install target is a Swift toolchain.

In conjunction with the change to swift/utils/build-script-impl
in PR#1212, this change enables Swift programs compiled using
the resulting built toolchain to successfully import
Dispatch with only -Xcc -fblocks as extra arguments to swiftc
(no longer need to also specify -I<path-to-dispatch-module>).