Switch libpwq submodule to upstream repository
Switch to mheily repository (all the needed changes from
dgrove-oss have been merged upstream). Tweak libdispatch's
configure.ac so that if we are building our own libpwq
from source, the primary configure passes down an argument to
libpwq's configure to modify libpwq's install targets
so that libpthread_workqueue.la is built as a libtool
convenience library and libpwg's man pages and header
files are not installed when libdispatch is installed.