resolve all compiler warnings
Resolve all compiler warnings when building libdispatch and
the test suite on linux.  It now compiles cleanly.

Many of the warnings were due to platform-specific sized integers
being used in printf-like contexts.  In libdispatch itself, I took
the approach of adding casts to the values to match the size/type
specified in the printf format string (under the assumption that
the expected output format should not be changed and needs to be
the same on all platforms).