Build fix for Dispatch overlay on Linux with 03242016 compiler.
The types mode_t and off_t are conditionally defined in
multiple header files in GLibc.  This interacts poorly with
the clang module maps, resulting in it insisting that these types
should be imported from stdio.h (which we don't want to
unconditionally include in the dispatch API header files).
We can fix the clash with mode_t by including sys/types.h
early in the list of includes in dispatch.h.  I was unable to
find a similar fix for off_t, so instead this change
conditionally includes stdio.h in dispatch/io.h under a
preprocessor symbol that will only be defined when compiling
Dispatch.swift to build the swiftmodule for the overlay.

Signed-off-by: Daniel A. Steffen <dsteffen@apple.com>