relax passing conditions for dispatch_concur
Further relax the expected minimum level of concurrency
when running on a multi-CPU system to only require that
we observe at least 50% of the potential concurrency is
actually realized.  This should be loose enough to avoid
spurious failures when run in the CI environment while
still being able to detect when we are very badly missing
the expected levels of concurrency.