adjust dispatch_timer_short to control testing time
Reduce N (number of suspended timers) from 100000 to 25000
to make it suitable for running in CI environment.
Test case execution time on Linux appears to be roughly quadratic
in N.  Setting N to 25000 yields a test that runs in about 30
seconds; at 100000 the test takes around 15 minutes to run.

Profiling says that 99% of test execution is spent in
_dispatch_timers_update.