Add dependency on swiftc for %.o.%.swift rules
Detect that swiftc has changed since the Swift file has
been compiled and force it to be recompiled.  The explicit
rm of the target before regenerating it is needed because if
the new compiler generates exactly the same bits as the old
one the .o file isn't updated and the target gets stuck as
being out-of-date.  An explicit rm avoids that problem.

Signed-off-by: Daniel A. Steffen <dsteffen@apple.com>