Remove unused include of sys/user.h
The include of sys/user.h is not needed; it
is a leftover from the initial linux port.
Eliminating it avoids a build problem for
the Swift dispatch overlay on ubuntu 14.04
powerpc64le because the content of sys/user.h
confuses the clang module importer when it
attempts to import CDispatch.