linux: update header used for `major` macro
Newer versions of glibc indicate that they intend to move the `major`
macro from `sys/types.h` to `sys/sysmacros.h`.  Add a check for the
header and include that earlier to ensure that the macro is provided by
the newer header when available/possible.  This avoids an unnecessary
warning from the system headers.