build: add support to build with sanitization
Add support to configure the build to use sanitizers to aid in finding
issues.  The `DISPATCH_USE_SANIZITERS` is a cmake option that defaults
to none, but can be one of `Address, Memory, MemoryWithOrigins,
Undefined, Leaks, Address;Undefined`.  This allows for easily building
libdispatch and the tests with the sanitizers.

Signed-off-by: Daniel A. Steffen <dsteffen@apple.com>