unify the assembly label by using the compiler
Rather than trying to enumerate the user label prefix on the various
targets, rely on the compiler.  The GNU compilers define
`__USER_LABEL_PREFIX__` to "_" as appropriate.  In the case that you are
using a non-GNU compiler, it is possible to specify the macro according
to your target.  This simplifies and unifies the code.