BlocksRuntime: adjust implementation for Windows x64
This adjust the BlocksRuntime to work on Windows x64 which is an LLP64
environment.  This primarily comes down to using `size_t` for the size
parameter for the allocation routines.  A secondary change which is
needed is the type cast for the `InterlockedCompareExchange` operation
on Windows.  A small cleanup is included which removes the definition of
an unused function.

Signed-off-by: Daniel A. Steffen <dsteffen@apple.com>