revive the Windows port
The `DISPATCH_EXPORT` macro added an extraneous `extern` on the
declaration.  This was pointed out by a newer clang.  Adjust the
declaration accordingly.

Add build rules to improve Windows builds.  These additions help
cross-compile to Windows with both clang-cl as well as the GNU clang
driver.

Use the newer Windows APIs to detect the windows CPU state.  This allows
us to collect all of the CPU configuration information to properly
affinitise work to the preferred CPU.

Use the FLS API to create thread local data storage for libdispatch's
queues.