workaround `-Wassign-enum` warnings
There are a number of cases where we use `0` or macro definitions for
values to enumerated types.  When building for Windows, these cause a
warning due to the value not being part of the enumeration.  This change
silences the warnings and identifies the various places where this
occurs.  Ideally, the values would be part of the enumerations.