build: silence `-Wassign-enum`
Silence the enum assignment warning from clang.  It is triggered by use
of constants in macros that are not part of an enumeration when the
parameter is of an enumerated type.  This shows up when building for
Windows.  Rather than clutter the codebase with compiler specific macros
to silence the warning, disable it.

Signed-off-by: Daniel A. Steffen <dsteffen@apple.com>