Requeue the stream handler if the read would block
There is a chance that the `read` operation could return a `EAGAIN` or
`EWOULDBLOCK` because the file descriptor used in the read operation of
`_dispatch_operation_perform` has been marked with `O_NONBLOCK`.

The existing code assumed that `EAGAIN` or `EWOULDBLOCK` would only happen when
there was no existing `dispatch_source_t`. Because of this assumption, it would
call `_dispatch_stream_source` which would create a `dispatch_source_t` and
register an event handler.

The bug occurs when `EAGAIN` or `EWOULDBLOCK` is returned from `read` when a
`dispatch_source_t` has already been allocated. This patch, explicitly watches
for the the described condition where the `dispatch_source_t` was already
created and then explicitly queues another call to
`_dispatch_stream_queue_handler` which re-attempts the `read` operation.

Fixes SR-5773.