Report crash messages on Windows to the CRT
Provide a Windows implementation of the _dispatch_set_crash_log_* macros which
pass the crash message to the CRT using _RPTF0. If libdispatch is linked against
a debug version of msvcrt, this will display a dialog with detailed information
about the crash and provide an option to attach a debugger. On release versions
of msvcrt, this reporting is a no-op.