tests: read from a temporary file instead of vi on non-Apple platforms
Some tests hardcode `/usr/bin/vi` as a path to a file to read from. This is
definitely not going to exist on platforms like Android. On non-Apple platforms,
instead of hardcoding this path in each test, generate a temporary file which is
sufficiently large and filled with random data.