build: always copy the modulemaps
Change from a symlink to a copy.  This is more portable and fixes the
distribution aspect.  When the installation occurs, the symbolic link is
not followed and a symbolic link is installed.  Rather if we copy, we
can get the contents.  The files are small so the cost is relatively
low.

Signed-off-by: Kim Topley <ktopley@apple.com>