Add string.h for memcpy in Data.swift.
In Data.swift the C function memcpy is used as part of _copyBytesHelper.
The function is defined in string.h, but that header is not referenced
directly by any of the headers in dispatch/ or by DispatchOverlayShims.h
in the compiler, which are the only two headers imported by the swift
file.

For some reason, this is working on Darwin and Ubuntu, probably because
some included header includes string.h or defines memcpy, avoiding the
problem. For those platforms that this already works, adding an extra
include should not affect them.