dispatch: include `time.h` on windows
time.h is needed on Windows to ensure that we have a definition of
`timespec`.  This is needed when building the swift SDK overlay.  It
seems that when cross-compiling, we were getting lucky with the
`timespec` definition being provided by some other header being
included.  This repairs the build of libdispatch's swift SDK overlay on
Windows targeting Windows.