dispatch: fix logging to a file on Windows
Port the dispatch_debug test to Windows and fix a bug that it found related to
logfile initialization.

_dispatch_logv_init()'s Windows implementation uses _fdopen() and fprintf() to
write the log header. However, this gives ownership of the file to stdio, so
fclose() closes the descriptor and causes future writes to trigger CRT assertion
failures. Avoid this by using snprintf() and _write() instead.