tests: port dispatch_io and dispatch_io_pipe_close to Windows
This lets us build and run the full extended test suite for Windows! I
caught several bugs in Dispatch's Windows port while I was working on
this and I submitted pull requests to fix them (#493, #494). The only
dispatch_io test case which fails right now is test_io_stop, and that's
because we don't have pipe support on Windows yet.