build: add DT_SONAME for ELF libraries
ELF libraries should have DT_SONAME set so that they can be loaded
properly.  This is needed currently since we do not use proper CMake
support for Swift and that requires that we pass along the `-soname`
flag to the linker ourselves.  Account for that.  This improves loading
of swiftDispatch on android.

Signed-off-by: Kim Topley <ktopley@apple.com>