event: implement source muxing on Windows
Implement the cases in `_dispatch_unote_register_muxed()` and
`_dispatch_unote_unregister_muxed()` for when multiple event sources are
open on a handle and we need to combine them.

The test suite doesn't hit these codepaths anywhere and we haven't run
into issues with Foundation yet, so I added a dispatch_io_muxed test
which opens multiple sources on a file/pipe/socket and checks that
events fire correctly.

Signed-off-by: Rokhini Prabhu <rokhinip@apple.com>