hw_config: fix Windows CPU detection
The Windows CPU detection code is not reporting the intended values.
`physical_cpus` should report the number of physical cores instead of
the number of packages, and `active_cpus` should report the number of
available logical cores instead of the number of available physical
cores. This fixes an issue where the `dispatch_apply` test times out on
hyperthreaded systems.

Signed-off-by: Rokhini Prabhu <rokhinip@apple.com>