queue: use the default stack size on Windows
The Windows thread pool implementation is calling `_beginthreadex` with
a stack size of 64 KB. This seems like a misunderstanding to me. In the
pthread code, `_dispatch_mgr_root_queue_init` does call
`pthread_attr_setstacksize` with a 64 KB size. However, this only
applies to the pthread root queue manager, and pthread root queues are
only supported on Apple platforms anyway.

We can confirm that using the OS default stack size is the intended
semantic by looking at the `dispatch_select` test, which expects to be
able to stack-allocate a 500 KB buffer in an event handler. This test is
failing on Windows because there is not enough stack space. We've also
run into various related issues in Swift which only happen on Windows,
and I suspect that this is just because the stack size isn't so small on
other platforms.