build: add exports targets
Now that we can build the Swift content as a library, we can do export
targets.  This enables us to have CMake determine more information about
the target, track dependencies across projects, and manage the flags
across the trees without the special variables to specify the source and
build trees.

Signed-off-by: Rokhini Prabhu <rokhinip@apple.com>