[build] Silence `-Wvoid-pointer-to-int-cast`
libdispatch will fail to build with a newly introduced clang diagnostic
`pointer-to-int-cast`. libdispatch converts a void pointer to a
dispatch_invoke_flags_t (aka unsigned int) in a few places. Ideally we would not
be doing this, but this solution at least gets libdispatch building again
with a newer version of clang.