Add a runloop implementation based on pipe2.
The Linux implementation uses an eventfd to impement the runloop
contract for CF. However, ordinary POSIX pipes are a widely available
API and can be used to achieve the same result as eventfd on other
systems that support it.

Since POSIX pipes require a pair of file descriptors, we force file
descriptors to 32-bit integers and pack these two into a single 64-bit
integer. This keeps the management of the runloop handle simple rather
than having to manage the handle storage externally, for the cost of
limiting the range of the file descriptor type.