Install Dispatch Exports in the Install Tree
When building foundation (or any project that depends on swiftDispatch)
using a built and installed toolchain that already contains libdispatch,
setting Dispatch_DIR to the source directory creates duplicate
definitions of the libdispatch headers.

This creates an additional dispatchConfig that gets installed to the
install tree that points to the installed headers rather than the source
headers.