Handle dispatch_unote_ident_t typedefs properly.
Different platforms need to typedef dispatch_unote_ident differently,
and subsequently may have different bit widths and proper handling
thereof. Mostly this involves debug printing, which needs proper format
handling.

One special case for consideration is `_dispatch_timer_unote_disarm`;
this gets the `du_ident` and converts that into a `uint32_t` timer
index. When `dispatch_unote_ident` is a `uint32_t` this is of course
fine, but needs consideration for when it isn't. Here, we just cast
down. This is somewhat reasonable since this is initialized from
`_dispatch_timer_unote_idx` which returns an `unsigned int`. (Of course,
we are not actually bounds checking that index, but that's outside the
scope of this commit).