Fix unused variable 'to' build failure
In only one of the ifdef cases is the `to` variable used.
Clang on the rebranch branch recognizes this to be unused on the other
branches and emits a warning. This project has `-Werror` enabled, so
this results in a build failure.

I moved the entire code block down closer to where it is used because it
isn't entirely clear based on where it used to sit given the number of
macro ifdef's between declaration and usage.