shims: convert `_dispatch_preemption_yeild` to an expression
Because `_dispatch_preemption_yield` is used an expression as:

~~~
((void) (q), _dispatch_preemption_yield(n))
~~~

we need to convert the expansion to an expression.  We _could_ also just
wrap the invocation into a parenthesis as an alternative approach, but
this should be safe enough to let us get away with it.